# Replace the default wasm garbage collector with nottinygc