--- a/scheduler/test/testdata.go
+++ b/scheduler/test/testdata.go
@@ -1,4 +1,4 @@
-package testdata
+package test
 
 import (
 	"bufio"
@@ -27,6 +27,10 @@ func NewPluginExampleFilterSimple(ctx context.Context) (frameworkruntime.Plugin,
 	})
 }
 
+var PathErrorPanicOnFilter = pathError("panic_on_filter")
+
+var PathTestPanicOnStart = pathError("panic_on_start")
+
 var PathExampleFilterSimple = pathExample("filter-simple")
 
 // NewPluginExampleNoop returns a new plugin configured with PathExampleNoop.
@@ -39,7 +43,7 @@ func NewPluginExampleNoop(ctx context.Context) (frameworkruntime.Plugin, error) 
 
 var PathExampleNoop = pathExample("noop")
 
-//go:embed testdata/node.yaml
+//go:embed testdata/yaml/node.yaml
 var yamlNodeReal string
 
 // NodeReal is a realistic v1.Node used for testing and benchmarks.
@@ -52,7 +56,7 @@ var NodeReal = func() *v1.Node {
 // NodeSmall is the smallest node that works with PathExampleFilterSimple.
 var NodeSmall = &v1.Node{ObjectMeta: apimeta.ObjectMeta{Name: "good-node"}}
 
-//go:embed testdata/pod.yaml
+//go:embed testdata/yaml/pod.yaml
 var yamlPodReal string
 
 // PodReal is a realistic v1.Pod used for testing and benchmarks.
@@ -89,11 +93,21 @@ func decodeYaml[O apiruntime.Object](yaml string, object O) {
 
 // pathExample gets the absolute path to a given example.
 func pathExample(name string) string {
+	return relativePath(path.Join("..", "..", "examples", name, "main.wasm"))
+}
+
+// pathError gets the absolute path wasm compiled from a %.wat source.
+func pathError(name string) string {
+	return relativePath(path.Join("testdata", "error", name+".wasm"))
+}
+
+// relativePath gets the absolute from this file.
+func relativePath(fromThisFile string) string {
 	_, thisFile, _, ok := runtime.Caller(1)
 	if !ok {
 		panic("cannot determine current path")
 	}
-	p := path.Join(path.Dir(thisFile), "..", "..", "examples", name, "main.wasm")
+	p := path.Join(path.Dir(thisFile), fromThisFile)
 	if abs, err := filepath.Abs(p); err != nil {
 		panic(err)
 		return ""