#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

After profiling in #24, I noticed over half the time in wasm was spent in garbage collection. This could have been guessed, but it was nice to verify it with [wzprof](https://github.com/stealthrocket/wzprof).

This uses an optimized garbage collector, [nottinygc](https://github.com/wasilibs/nottinygc), dramatically improving performance.

#### Which issue(s) this PR fixes:

NONE

#### Special notes for your reviewer:

https://github.com/wasilibs/nottinygc was invented by the excellent @anuraaga

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

same code takes half the time!

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
                                            │  before.txt   │             after.txt              │
                                            │    sec/op     │   sec/op     vs base               │
PluginFilter/noop/params:_small-12             148.6n ±  2%   143.8n ± 1%   -3.20% (p=0.002 n=6)
PluginFilter/noop/params:_real-12              146.5n ±  0%   143.2n ± 0%   -2.25% (p=0.002 n=6)
PluginFilter/filter-simple/params:_small-12   14.436µ ±  5%   7.234µ ± 2%  -49.89% (p=0.002 n=6)
PluginFilter/filter-simple/params:_real-12     232.5µ ± 21%   121.1µ ± 1%  -47.91% (p=0.002 n=6)
geomean                                        2.923µ         2.061µ       -29.50%
```