@codefromthecrypt 2023-05-31
I should confess one downside, which is that the wasm is significantly larger (+110KB) However, the performance is quite bad and I think this is an easy tradeoff.