@anuraaga 2023-06-01
Updated `nottinygc` to print a grokkable message when the TinyGo flags haven't been set

https://github.com/wasilibs/nottinygc/releases/tag/v0.3.0