@codefromthecrypt 2023-06-01
Thanks @anuraaga I updated like so

```
$ for I in `find . -name go.mod -exec egrep -q nottinygc {} \; -print`; do (cd `dirname $I`; go get  github.com/wasilibs/nottinygc@v0.3.0; go mod tidy); done
```