@codefromthecrypt 2023-06-01
so here's the current error if the plugin wasn't built with the flags:
```
    plugin_perf_test.go:38: failed to create plugin: wasm: error instantiating guest: module[filter-simple-1] function[_start] failed: wasm error: unreachable
        wasm stack trace:
        	.runtime._panic(i32,i32)
        	._start.command_export()
```