@codefromthecrypt 2023-06-01
>  would require stderr/stdout to be available

this might be ok. lemme wire up and see