@sanposhiho 2023-06-03
@codefromthecrypt 

> I should confess one downside, which is that the wasm is significantly larger (+110KB) However, the performance was quite bad and I think this is an easy tradeoff.
https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/pull/25#issuecomment-1569889449

In the server-side wasm, what gonna be the problem/disadvantage of a bigger wasm size? The performance to load the wasm in the host side?