@salaboy 2023-06-03
Yeah configMaps have size restrictions, but also you are storing those
files in etcd, which was not designed for storing files (in the context of
kubernetes) , saving wasm files in volumes is the way to go

On Sat, 3 Jun 2023 at 06:59, Crypt Keeper ***@***.***> wrote:

> In the server-side wasm, what gonna be the problem/disadvantage of a
> bigger wasm size? The performance to load the wasm in the host side?
>
> One reason is I think some people move wasm around with config maps? cc
> @salaboy <https://github.com/salaboy> about any size constraints there,
> in case folks here aren't aware of the nuance. Basically, we may be 110kb
> closer to that limit. Once past that limit, we need to use http refs like
> most proxy-wasm are moved around by istio (OCI is another way, such as what
> trivy does).
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/pull/25#issuecomment-1574658951>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACCMXXVQIG4DWIDDSNM4DLXJLAERANCNFSM6AAAAAAYVGRKSE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***
> com>
>
-- 
 - Blog: http://salaboy.com <http://salaboy.wordpress.com>
 - Github user: http://github.com/salaboy
 - Twitter: http://twitter.com/salaboy

 - Mauricio "Salaboy" Salatino -