@codefromthecrypt 2023-06-04
I'll go ahead and add the unit test here and also add an issue for docs on how to handle wasm distribution