@codefromthecrypt 2023-06-04
ok I backfilled tests for panic messages PTAL!