@codefromthecrypt 2023-06-04
added a `make testdata` target which makes all the test wasm