@codefromthecrypt 2023-06-05
I will add a RATIONALE section then ping @sanposhiho for final review